@@ -1,40 +1,43 @@
 {
-    "name": "mvc-tech-blog",
-    "version": "1.0.0",
-    "description": "",
-    "main": "server.js",
-    "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1",
-      "start": "node server.js",
-      "seed": "node seeds/seed.js"
-    },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-      "bcrypt": "^5.0.0",
-      "connect-session-sequelize": "^7.0.4",
-      "dotenv": "^8.2.0",
-      "express": "^4.17.1",
-      "express-handlebars": "^5.2.0",
-      "express-session": "^1.17.1",
-      "mysql2": "^2.2.5",
-      "sequelize": "^6.28.0"
-    },
-    "devDependencies": {
-      "@babel/eslint-parser": "^7.19.1",
-      "@babel/plugin-syntax-import-assertions": "^7.20.0",
-      "autoprefixer": "^10.4.13",
-      "dependency-cruiser": "^12.5.1",
-      "eslint": "^8.32.0",
-      "eslint-config-prettier": "^8.6.0",
-      "eslint-config-standard": "^17.0.0",
-      "eslint-config-xo": "^0.43.1",
-      "eslint-plugin-import": "^2.27.5",
-      "eslint-plugin-n": "^15.6.1",
-      "eslint-plugin-promise": "^6.1.1",
-      "postcss": "^8.4.21",
-      "prettier": "^2.7.1",
-      "tailwindcss": "^3.2.4"
-    }
+  "name": "social-book-club",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js",
+    "watch": "nodemon server.js",
+    "seed": "node seeds/seed.js"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "bcrypt": "^5.0.0",
+    "connect-session-sequelize": "^7.0.4",
+    "dotenv": "^8.6.0",
+    "express": "^4.18.2",
+    "express-handlebars": "^5.3.5",
+    "express-session": "^1.17.3",
+    "flowbite": "^1.6.3",
+    "handlebars": "^4.7.7",
+    "mysql2": "^2.3.3",
+    "sequelize": "^6.28.0"
+  },
+  "devDependencies": {
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-syntax-import-assertions": "^7.20.0",
+    "dependency-cruiser": "^12.5.1",
+    "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-config-xo": "^0.43.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "nodemon": "^2.0.15",
+    "prettier": "^2.8.3",
+    "tailwindcss": "^3.2.4",
+    "vitest": "^0.28.1"
   }
+}
