@@ -1,21 +1,20 @@
 require("dotenv").config();
-const Sequelize = require("sequelize");
+import Sequelize from "sequelize";
 
-let sequelize;
-
-if (process.env.JAWSDB_URL) {
-    sequelize = new Sequelize(process.env.JAWSDB_URL);
-} else {
-    sequelize = new Sequelize(
-        process.env.DB_NAME,
-        process.env.DB_USER,
-        process.env.DB_PASSWORD,
-        {
-            host: "127.0.0.1",
-            dialect: "mysql",
-            port: 3306,
-        }
+// create connection to our db
+const sequelize = process.env.JAWSDB_URL
+  ? // Here ? is a placeholder for the value of process.env.JAWSDB_URL
+    new Sequelize(process.env.JAWSDB_URL)
+  : // And : is a placeholder for the value of the object
+    new Sequelize(
+      process.env.DB_NAME,
+      process.env.DB_USER,
+      process.env.DB_PASSWORD,
+      {
+        host: "127.0.0.1",
+        dialect: "mysql",
+        port: 3306,
+      }
     );
-}
 
-module.exports = sequelize;
+export default sequelize;
