@@ -1,5 +1,5 @@
-const { Sequelize, Model, DataTypes } = require("sequelize");
-const sequelize = require("../config/connection.js");
+import { Sequelize, Model, DataTypes } from "sequelize";
+import sequelize from "../config/connection.js";
 
 // create our Comment model
 class Comment extends Model {}
@@ -18,4 +18,4 @@ Comment.init(
 );
 
 // export the Comment model
-module.exports = Comment;
+export default Comment;
