@@ -1,12 +1,12 @@
-const { Sequelize, Model, DataTypes } = require("sequelize");
-const sequelize = require("../config/connection.js");
-const bcrypt = require("bcrypt");
+import { Model, DataTypes } from "sequelize";
+import sequelize from "../config/connection.js";
+import { compareSync, hash } from "bcrypt";
 
 // create our User model
 class User extends Model {
     // set up method to run on instance data (per user) to check password
     checkPassword(loginPw) {
-        return bcrypt.compareSync(loginPw, this.password);
+        return compareSync(loginPw, this.password);
     }
 }
 
@@ -36,17 +36,13 @@ User.init(
         hooks: {
             // set up beforeCreate lifecycle "hook" functionality
             async beforeCreate(newUserData) {
-                newUserData.password = await bcrypt.hash(newUserData.password, 10);
+                newUserData.password = await hash(newUserData.password, 10);
                 return newUserData;
             },
             // set up beforeUpdate lifecycle "hook" functionality
             async beforeUpdate(updatedUserData) {
-                updatedUserData.password = await bcrypt.hash(updatedUserData.password, 10);
+                updatedUserData.password = await hash(updatedUserData.password, 10);
                 return updatedUserData;
-            },
-            async beforeBulkCreate(newUserData) {
-                newUserData.password = await bcrypt.hash(newUserData.password, 10);
-                return newUserData;
             },
         },
         sequelize,
@@ -58,4 +54,4 @@ User.init(
 );
 
 // export the User model
-module.exports = User;
+export default User;
