@@ -1,6 +1,6 @@
-const Comment = require("./Comment.js");
-const Post = require("./Post.js");
-const User = require("./User.js");
+import Comment from "./Comment.js";
+import { belongsTo, hasMany } from "./Post.js";
+import User from "./User.js";
 
 // create associations
 Comment.belongsTo(User, {
@@ -8,12 +8,12 @@ Comment.belongsTo(User, {
     onDelete: "CASCADE",
 });
 
-Post.belongsTo(User, {
+belongsTo(User, {
     foreignKey: "user_id",
     onDelete: "CASCADE",
 });
 
-Post.hasMany(Comment, {
+hasMany(Comment, {
     foreignKey: "post_id",
     onDelete: "CASCADE",
 });
